+++ src/components/DataVisualization/DataTable/DataTable.stories.mdx
@@ -86,8 +86,7 @@ export const Template = (args, { argTypes }) => ({
         :title="title"
         :entities="entities"
         :entitySchema="entitySchema"
-        :showHeader="showHeader"
-        :showSortIcon="showSortIcon"
+        :settings="{showSortIcon, showHeader}"
       >
         <template #actions>
           <PbButton
